pyproject.toml
[tool.poetry]
name = "griptape"
version = "0.20.1"
description = "Modular Python framework for LLM workflows, tools, memory, and data."
authors = ["Griptape <hello@griptape.ai>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/griptape-ai/griptape"

packages = [
    {include = "griptape"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
openai = "^1.1.1"
attrs = ">=22"
jinja2 = ">=3.1"
jsonschema = ">=4"
marshmallow = ">=3"
marshmallow-enum = ">=1.5"
tiktoken = ">=0.3"
rich = ">=13"
schema = ">=0.7"
pyyaml = ">=6"
tenacity = ">=8.0"
numpy = ">=1"
PyPDF2 = ">=3"
trafilatura = ">= 1.6"
pandas = ">=1.3"
stringcase = "^1.2.0"
docker = "^6.1.3"
sqlalchemy = "~=1.0"
cryptography = "^41.0.4"
dateparser = "^1.1.8"
mail-parser = "^3.15.0"
requests = "^2"


cohere = { version = ">=4", optional = true }
anthropic = { version = "^0.3.0", optional = true }
transformers = { version = "^4.30", optional = true }
huggingface-hub = { version = ">=0.13", optional = true }
boto3 = { version = "^1.28.2", optional = true }
sqlalchemy-redshift = { version = "*", optional = true }
snowflake-sqlalchemy = { version = "^1.4.7", optional = true }
pinecone-client = { version = ">=2", optional = true }
pymongo = { version = "*", optional = true }
marqo = { version = ">=1.1.0", optional = true }
redis = { version = "^4.6.0", optional = true }
opensearch-py = { version = "^2.3.1", optional = true }
requests-aws4auth = { version = "1.2.3", optional = true }
pgvector = { version = "^0.2.3", optional = true }
psycopg2-binary = { version = "^2.9.9", optional = true }
pillow = "^10.1.0"


[tool.poetry.extras]
drivers-prompt-cohere = ["cohere"]
drivers-prompt-anthropic = ["anthropic"]
drivers-prompt-huggingface = ["huggingface-hub", "transformers"]
drivers-prompt-amazon-bedrock = ["boto3", "anthropic"]
drivers-prompt-amazon-sagemaker = ["boto3", "transformers"]

drivers-sql-redshift = ["sqlalchemy-redshift", "boto3"]
drivers-sql-snowflake = ["snowflake-sqlalchemy", "snowflake", "snowflake-connector-python"]
drivers-sql-postgres = ["pgvector", "psycopg2-binary"]

drivers-memory-conversation-amazon-dynamodb = ["boto3"]

drivers-vector-marqo = ["marqo"]
drivers-vector-pinecone = ["pinecone-client"]
drivers-vector-mongodb = ["pymongo"]
drivers-vector-redis = ["redis"]
drivers-vector-opensearch = ["opensearch-py"]
drivers-vector-amazon-opensearch = ["opensearch-py", "boto3", "requests-aws4auth"]
drivers-vector-postgresql = ["pgvector", "psycopg2-binary"]

all = [
    "cohere",
    "anthropic",
    "hugginface-hub",
    "transformers",
    "sqlalchemy-redshift",
    "boto3",
    "snowflake-sqlalchemy",
    "snowflake",
    "marqo",
    "pinecone-client",
    "pymongo",
    "redis",
    "opensearch-py",
    "requests-aws4auth",
    "pgvector",
    "psycopg2-binary",
]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "~=7.1"
pytest-mock = "*"
mongomock = "*"

twine = ">=4"
moto = {extras = ["dynamodb"], version = "^4.1.8"}
pytest-xdist = "^3.3.1"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.1"
fuzzywuzzy = "^0.18.0"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.286"
black = "^23.7.0"
pyright = "^1.1.324"
pre-commit = "^3.5.0"

[tool.black]
line-length=120
skip_magic_trailing_comma = true

[tool.ruff]
line-length=120

[tool.pyright]
include = ["griptape"]
exclude = [
    "**/__pycache__",
]
pythonVersion = "3.9"
reportOptionalMemberAccess = "none"

[tool.pytest_env]
OPENAI_API_KEY = {value = "api-key", skip_if_set = true}
AZURE_OPENAI_API_KEY = { value = "api-key", skip_if_set = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
